requirements.txt
Django==4.1.7
mysqlclient==2.1.1
django-enumfields
